backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
faiss-cpu>=1.9.0
sentence-transformers>=2.7.0
pydantic>=2.6.1
python-dotenv>=1.0.1
numpy>=1.26.0
pyodbc>=5.0.0
pymssql>=2.3.0
sqlalchemy>=2.0.0
